Separated GT model page and upaded monaco editor
